storage-service/requirements.txt
uvicorn 
fastapi
pika
python-dotenv
python-multipart